dump: print_return_trace: Fix printing of filename
1.
file_path() from the kernel fills the path buffer in reverse order from
the end to the beginning of the buffer so we need to walk the buffer
backwards

2. Detect if we're returning to user space or to the kernel.

TODO:

1. For the kernel driver. When tracing starts the driver needs to walk
all processes (for_each_process()) and extract all the existing maps.
Currently we will end up with a 'none' filename for all existing processes.

2. For vmlinux we need to extract the start address somehow and add that
to the pc.

3. Print which function??? Use libelf or not???