dump: Handle 'incorrect' traces
This is kind of a hack for now.

Added status flags to the H/W FIFO (full & half) and they never get
triggered so I don't think any traces get dropped from the stream.

We once in a while get an extra spurious(?) RETURN trace that seemingly
pushes the frame level up above the top callsite.

For now in that case we just pretend that never happened and set the
current frame level back to 0 (top).

My best guess is that something weird happens when a task gets migrated
to another CPU. Perhaps we also need to track sched_in events too in the
kernel driver ...